move fixtures to own module
